backend/requirements.txt
librosa
scipy
numpy
flask  # or fastapi
sqlalchemy
yt-dlp
spotipy
python-socketio
requests
pytest
sqlalchemy
python-dotenv